requirements.txt
langgraph
langchain-google-genai
python-dotenv
langchain
lark
fastapi
uvicorn[standard]
requests
